ftl_travel_mcp_server/requirements.txt
mcp==1.26.0
python-dotenv==1.1.1
requests==2.32.5
fastmcp==2.14.5
mcp==1.26.0
asyncio==4.0.0